requirements.txt
matplotlib==3.6.2
numpy==1.23.3
Pillow==10.0.0
requests==2.28.1
streamlit==1.23.1
tqdm==4.64.1
